+++ src/main/java/ev3dev/sensors/slamtec/RPLidarA1Fake.java
@@ -29,7 +29,7 @@ import java.util.concurrent.atomic.AtomicInteger;
     public Scan scan() throws RPLidarA1ServiceException {
         final List<ScanDistance> distances = Collections.synchronizedList(new ArrayList<>());
         for(int angle = 0; angle < 360; angle++){
-            distances.add(new ScanDistance(angle,Math.random() * 4000 + 1, 1, false));
+            distances.add(new ScanDistance(angle, new Float(Math.random() * 4000 + 1), 1, false));
         }
         return new Scan(distances);
     }