+++ src/main/java/ev3dev/sensors/slamtec/service/RpLidarLowLevelDriver.java
@@ -33,6 +33,7 @@ public @Slf4j class RpLidarLowLevelDriver {
 	public static final byte RCV_INFO = (byte) 0x04;
 	public static final byte RCV_HEALTH = (byte) 0x06;
 	public static final byte RCV_SCAN = (byte) 0x81;
+	private static final byte START_MOTOR = (byte) 0xF0;
 
 	SerialPort serialPort;
 	InputStream in;
@@ -197,6 +198,64 @@ public @Slf4j class RpLidarLowLevelDriver {
 			//TODO Think in a Low Level exception
 			e.printStackTrace();
 		}
+	}
+
+	/**
+	 * Sends a command with data payload
+	 */
+	protected void sendPayLoad(byte command, byte[] payLoad) {
+		if (verbose) {
+			System.out.printf("Sending command 0x%02x\n", command & 0xFF);
+		}
+
+		dataOut[0] = SYNC_BYTE0;
+		dataOut[1] = command;
+
+		//add payLoad and calculate checksum
+		dataOut[2] = (byte) payLoad.length;
+		int checksum = 0 ^ dataOut[0] ^ dataOut[1] ^ (dataOut[2] & 0xFF);
+
+		for(int i=0; i<payLoad.length; i++){
+			dataOut[3 + i] = payLoad[i];
+			checksum ^= dataOut[3 + i];
+		}
+
+		//add checksum - now total length is 3 + payLoad.length + 1
+		dataOut[3 + payLoad.length] = (byte) checksum;
+
+		try {
+			out.write(dataOut, 0, 3 + payLoad.length + 1);
+			out.flush();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+	}
+
+	/**
+	 * Sends a command with data payload - int
+	 */
+	protected void sendPayLoad(byte command, int payLoadInt) {
+		byte[] payLoad = new byte[2];
+
+		//load payload little Endian
+		payLoad[0] = (byte) payLoadInt;
+		payLoad[1] = (byte) (payLoadInt >> 8);
+
+		sendPayLoad(command, payLoad);
+	}
+
+	/**
+	 * Sends a start motor command
+	 */
+	public void sendStartMotor(int speed) {
+		sendPayLoad(START_MOTOR, speed);
+	}
+
+	/**
+	 * Sends a stop motor command
+	 */
+	public void sendStopMotor() {
+		sendPayLoad(START_MOTOR, 0);
 	}
 
 	/**